+++ src/duplocloud/cli.py
@@ -1,8 +1,11 @@
+import sys
 from duplocloud.controller import DuploCtl
+from duplocloud.commander import resource_help_intercept
 from duplocloud.errors import DuploError
-from sys import exit
 
 def main():
+  if resource_help_intercept(sys.argv[1:]):
+    return
   try:
     duplo, args = DuploCtl.from_env()
     o = duplo(*args)
@@ -10,10 +13,10 @@ def main():
       print(o)
   except DuploError as e:
     print(e)
-    exit(e.code)
+    sys.exit(e.code)
   except Exception as e:
     print(f"An unexpected error occurred: {e}")
-    exit(1)
+    sys.exit(1)
 
 if __name__ == "__main__":
   main()